Move geometry constructor json classes to a sub-package
